QBITS_SkillMap_15/pcce-hackathon/backend/interview/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
python-multipart
httpx
pydantic>=2.0
supabase
ffmpeg-python
opencv-python-headless
tavily-python
PyGithub
pdfplumber
pytesseract
Pillow
weasyprint
python-docx
groq
beautifulsoup4
playwright